requirements.txt
asgiref==3.11.0
bcrypt==4.1.3
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.11
iniconfig==2.3.0
mysql-connector-python==9.0.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pytest==8.4.2
python-dotenv==1.0.1
requests==2.32.5
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
pyyaml
inflection
django-filter==22.1  
